refactor code structure
