+++ frontend/src/model/types.d.ts
