+++ frontend/src/pages/index.tsx
@@ -1,5 +1,5 @@
-export { Create } from './create/Create';
-export { Detail } from './detail/Detail';
-export { Edit } from './edit/Edit';
-export { Error } from './error/Error';
-export { Home } from './home/Home';
+export { CreatePage as Create } from './create/CreatePage';
+export { DetailPage as Detail } from './detail/DetailPage';
+export { EditPage as Edit } from './edit/EditPage';
+export { ErrorPage as Error } from './error/ErrorPage';
+export { HomePage as Home } from './home/HomePage';
