+++ frontend/src/router/AppRouter.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Navigate, Route, Routes } from 'react-router-dom';
-import { Create, Detail, Edit, Error, Home } from '../../pages';
-import { Layout } from '../../shared/components/layout/Layout';
+import { Create, Detail, Edit, Error, Home } from '../pages';
+import { Layout } from '../shared/components/layout/Layout';
 
 export default function AppRouter() {
   return (
