+++ frontend/src/shared/components/index.ts
@@ -1,2 +1,5 @@
 export { ErrorMessage } from './error';
 export { BookForm } from './form/BookForm';
+export { Layout } from './layout/Layout';
+export { Spinner } from './spinner/Spinner';
+export { Toast } from './toast/Toast';
