+++ frontend/src/utils/uuid.ts
@@ -1,12 +1,3 @@
-// API
-/**
- * Create an unique hash code
- * @returns string
- */
 export function uuidv4(): string {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-    const r = (Math.random() * 16) | 0;
-    const v = c === 'x' ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
+  return self.crypto.randomUUID();
 }
